Hybrid Scheme
This variation is a hybrid that uses the blank.gif, but also does
string manipulation to find the directory path, if necessary. It takes
into account a CMS scenario where the path to the blank.gif and the
responsive image may be different by adding an optional data attribute
for the directory path.